Fix IPv6 pools
The current method used by prometheus-openstack-exporter to find the
number of IP addresses in a pool is to have iter_iprange enumerate the
addresses in the pool. This works for small pools which are quite
typical for IPv4 address pools. With IPv6 pools, however, typical pool
sizes are large enough that prometheus-openstack-exporter will try to
eat up all of the computer's RAM which, depending on the host system's
configuration, can result in the host becoming completely unresponsive.

The value of showing the user the number of addresses in such a large
IPv6 pool is dubious to begin with, and adding it to the number of IPv4
addresses just ensures that the resulting number is virtually
meaningless.

Therefore, this patch simply skips IPv6 addresses when counting
IP addresses.