Fix regression that made all environment variables compulsory
The function, get_creds_list, has recently been fixed to ensure that
it does not return an array with elements in incorrect positions if
not all expected OS_* environment variables are present.

get_creds_dict received the same treatment. This was not necessary
because that function returns a dictionary: elements may be missing, but
one will never confused for another (because both key and value will be
missing). However, the change in get_creds_dict caused a problem: every
single argument must have the corresponding environment variable set or
the script will abort, forcing the user to provide redundant arguments
(e.g., PROJECT_DOMAIN_ID when PROJECT_DOMAIN_NAME is already present).

The patch reverts get_creds_dicts back to the correct version.