Disable setuptools package automatic discovery
setuptools-61.0.0 introduced automatic discovery for packages if no
`packages` or `py_modules` is explicitly specified. Package
autodiscovery causes Debian builds to fail with:

  error: Multiple top-level packages discovered in a flat-layout: ['snap', 'debian'].

More information about this bug can be found in:

  https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1021925

This can be fixed by disabling package autodiscovery, as I believe
that this software does not rely on that feature. More information in:

- https://setuptools.pypa.io/en/latest/history.html#v61-0-0
- https://github.com/pypa/setuptools/issues/3227
- https://github.com/pypa/setuptools/issues/3227#issuecomment-1081688343